requirements.txt
pandas
nba_api
schedule
pyyaml